src/dashboard/src/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=dev.txt dev.in
#
-e git+https://github.com/Brown-University-Library/django-shibboleth-remoteuser.git@67d270c65c201606fb86d548493d4b3fd8cc7a76#egg=django-shibboleth-remoteuser
-e git+https://github.com/seatme/django-longer-username.git@seatme#egg=longerusername
agentarchives==0.5.0
amclient==1.1.1
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backports.shutil-get-terminal-size==1.0.0  # via ipython
bagit==1.7.0
brotli==0.5.2
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
click==7.0                # via pip-tools
contextlib2==0.5.5
coverage==4.2
cryptography==2.6.1
decorator==4.4.0          # via ipython, traitlets
django-auth-ldap==1.3.0
django-autoslug==1.9.3
django-braces==1.0.0
django-extensions==1.1.1
django-forms-bootstrap==3.1.0
django-model-utils==1.3.1
django-prometheus==1.0.15
django-tastypie==0.13.2
django==1.8.19
elasticsearch==6.3.1
enum34==1.1.6
filelock==3.0.10
funcparserlib==0.3.6
funcsigs==1.0.2
future==0.17.1
futures==3.2.0 ; python_version < "3.0"
gearman==2.0.2
gevent==1.3.6
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
ipaddress==1.0.22
ipdb==0.12
ipython-genutils==0.2.0   # via traitlets
ipython==5.8.0
lazy-paged-sequence==0.3
logutils==0.3.3
lxml==3.5.0
metsrw==0.3.14
mock==2.0.0
mockldap==0.2.8
more-itertools==5.0.0
mysqlclient==1.3.7
ndg-httpsclient==0.5.1
pathlib2==2.3.3
pbr==5.1.3
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==4.3.0
pluggy==0.9.0
prometheus-client==0.7.1
prompt-toolkit==1.0.15    # via ipython
ptyprocess==0.6.0         # via pexpect
py==1.8.0
pyasn1-modules==0.2.5
pyasn1==0.4.5
pycparser==2.19
pygments==2.3.1           # via ipython
pyopenssl==19.0.0
pytest-cov==2.4.0
pytest-django==3.4.8
pytest-mock==2.0.0
pytest-pythonpath==0.7.3
pytest==4.3.1
python-dateutil==2.6.0
python-ldap==3.2.0
python-mimeparse==1.6.0
pytz==2018.9
pyyaml==5.1
requests==2.21.0
scandir==1.10.0
simplegeneric==0.8.1      # via ipython
six==1.12.0
toml==0.10.0
tox==3.8.3
traitlets==4.3.2          # via ipython
urllib3==1.24.1
vcrpy==1.13.0
virtualenv==16.4.3
wcwidth==0.1.7            # via prompt-toolkit
whitenoise==3.3.0
wrapt==1.11.1

# The following packages are considered to be unsafe in a requirements file:
# setuptools
